Update tab-unloading-v2-windows-release
Overall and weekly - Added new metrics to end under a ## line for both.  Let me know if the ## will work mid code block.  Trying to make it sane to proof read the changes.

Metrics sections - Added these to the END of the existing ones from previous tab unloading expeirment - in an attempt to make them easier to review.  Also did this because I didn’t want to go crazy trying to map the Perf ones together.  It may be harder to read the partyball results though.  :( Reviewer thoughts on how to optimize for both in experiments like this?

Data Source - Added missing data source (from the original tab unloading TOML versus the JS parsing TOML) to END of data source list.  How do we know if we even needed that data source?  Is it an issue to include a data source that we don’t end up using?